Revert "feat(calculator): blackhole spotlight focus"
This reverts commit 0ed49e7bb3eb5b37ede4a70d33561e978e1aac42.
